feat: 优化代码

+++ app/src/main/java/io/legado/app/ui/book/read/page/ContentTextView.kt
@@ -50,7 +50,7 @@ class ContentTextView : View {
                 textLine.textChars.forEach {
                     canvas.drawText(
                         it.charData,
-                        it.leftBottomPosition.x.toFloat(),
+                        it.leftBottomPosition.x,
                         it.leftBottomPosition.y.toFloat(),
                         textPaint
                     )
